Create a DocumentService
This sevice has been created to help reduce the size of scope of the `OffenderService`
